Added server.py
This is the main routine that generates an https page that uses
websockets to send blocks of audio to a server. I'm hoping I can
use this as the front end for the next version of Naomi. Because
it can either use a provided TLS certificate or generate one
itself, this application can be accessed remotely, including from
cell phones.